+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Unreleased changes
+
+* feature: Allow a `colspec_file` config with column info for `fixedwidth` inputs
+
 ### v0.4.1
 <details>
 <summary>Released 2024-11-15</summary>